all pages update, before report

+++ app/Http/Middleware/EnsureAdminRole.php
@@ -17,7 +17,7 @@ class EnsureAdminRole
     {
         $user = $request->user();
 
-        if ($user && $user->role !== 'superadmin' && $user->role !== 'admin') {
+        if ($user && $user->role !== 'superadmin' && $user->role !== 'admin' && $user->role !== 'bhw') {
             // Revoke the current token
             // $request->user()->currentAccessToken()->delete();
             $request->user()->tokens()->delete();
